+++ src/main/java/com/globalshopper/GlobalShopper/dto/response/ParticipationResponseDTO.java
@@ -12,7 +12,7 @@ public record ParticipationResponseDTO(
         CommercantResponseDTO commercantResponseDTO,
         LocalDate data,
         int quantite,
-        int montant,
+        double montant,
         Transaction transaction
 ) {
 }
