+++ src/main/java/com/globalshopper/GlobalShopper/dto/response/ProduitResponseDTO.java
@@ -9,7 +9,7 @@ public record ProduitResponseDTO(
         long id,
         String nom,
         String description,
-        int prix,
+        double prix,
         int moq,
         int stock,
         UniteProduit unite,
